Enables decimal calculations

+++ script.js
@@ -69,8 +69,8 @@ equalsButton.addEventListener("click", ()=>{
         operationsArr=inputString.split('/');
         operation="/";
     } ;
-    firstNumber=parseInt(operationsArr[0]);
-    secondNumber=parseInt(operationsArr[1]);
+    firstNumber=parseFloat(operationsArr[0]);
+    secondNumber=parseFloat(operationsArr[1]);
     result=operate(firstNumber, operation, secondNumber).toPrecision(12);
     display.textContent=result;
     inputString=result;
